Wake and sleep events with "push back" data
This module's watch_read function wakes and sleeps events using the
AnyEvent io event watcher API.  But in the case of synthetic reads
("push_back_read"), the "event_watch" bit must also be set, so that
the idle event (PostEventLoop) will honor the current setting.  So, in
the case of push back data, set or clear the POLLIN bit appropriately.